+++ lib/lib_display/JaretBurkett_ILI9488-gemu-1.0/ILI9488.h
@@ -141,7 +141,7 @@ class ILI9488 : public Renderer {
   void setScrollArea(uint16_t topFixedArea, uint16_t bottomFixedArea);
   void scroll(uint16_t pixels);
   void pushColor(uint16_t color);
-  void pushColors(uint16_t *data, uint8_t len, boolean first);
+  void pushColors(uint16_t *data, uint16_t len, boolean first);
   //void drawImage(const uint8_t* img, uint16_t x, uint16_t y, uint16_t w, uint16_t h);
   void fillScreen(uint16_t color);
   void drawPixel(int16_t x, int16_t y, uint16_t color);
